Introduce Auto-Format to Sprig Editor #1399 - Edited (#1504)
* Implemented Auto Format Code #1399

As requested on #1399, this is the first irritation of the main feature starting with basic beautify format with white-space

* Added Beginner Friendly Comments

As @JosiasAurel requested , i have added a beginner friendly comment on the function i wrote to make any new user read code base and understand it's uses.

* Rename Function Name For Better Understanding

* Update Format Option

after @leomcelroy feedback, i changed the current version with his which seems to be workable in the way we want.

* Hide Nav On Clicked

As requested by @leomcelroy i have made sure that navpoup closes / hide when the button is clicked.

* import entire js_beutify module before getting beutifier function (#1535)

* Collapse On Prettify

removed js-beautify comment and added a function that collapse the asset after it's formated as @leomcelroy suggsted

---------

Co-authored-by: Josias Aurel <ndjosiasaurel@gmail.com>